Remove parsed_results.rs
